feat: queryChartConfig method + events chart ratio
